Relies solely on env variables. Compares local to upstream and upstream to local. Split into 3 classes.
