+++ .env.example
@@ -1,0 +1,12 @@
+REPO_DIR=C:\\Instrument\\Apps\\EPICS\\
+UPSTREAM_BRANCH_CONFIG=epics
+
+WORKSPACE=temporary_workspace
+
+SSH_CREDENTIALS_USR=***REMOVED***
+SSH_CREDENTIALS_PSW=***REMOVED***
+
+USE_TEST_INSTRUMENT_LIST=false
+TEST_INSTRUMENT_LIST=NDXSCIDEMO
+
+DEBUG_MODE=false
