--- Jenkinsfile_CONFIG
@@ -1,59 +1,0 @@
-#!groovy
-
-pipeline {
-    // using same agnert as ConfigCheck job
-    agent {
-        label {
-            label 'ConfigCheck'
-        }
-    }
-
-    triggers {
-        cron('H 8 * * *')
-    }
-
-    environment {
-        SSH_CREDENTIALS = credentials('SSH')
-        TEST_INSTRUMENT_LIST = "${TEST_INSTRUMENT_LIST}"
-        USE_TEST_INSTRUMENT_LIST = "${USE_TEST_INSTRUMENT_LIST}"
-        DEBUG_MODE = "${DEBUG_MODE}"
-        REPO_DIR = "C:\\Instrument\\settings\\config\\common\\"
-        UPSTREAM_BRANCH = "master"
-    }
-
-    stages {
-        stage('Checkout') {
-            steps {
-                timeout(time: 2, unit: 'HOURS') {
-                    retry(5) {
-                        checkout scm
-                    }
-                }
-            }
-        }
-
-        stage('Check Instrument has any Hotfixes and then any uncommitteed changes') {
-            steps {
-                echo 'Check Instrument has any Hotfixes and then any uncommitteed changes'
-                timeout(time: 1, unit: 'HOURS') {
-                    bat '''
-                        call jenkins_utils/hotfix_checker.bat
-                    '''
-                }
-            }
-        }
-    }
-
-    post {
-        always {
-            archiveArtifacts artifacts: '*git_status/*.txt', caseSensitive: false
-            logParser([
-                projectRulePath: 'parse_rules',
-                parsingRulesPath: '',
-                showGraphs: true,
-                unstableOnWarning: true,
-                useProjectRule: true,
-            ])
-        }
-    }
-}
