+++ utils/hotfix_utils/check.py
@@ -1,0 +1,10 @@
+from enum import (
+    Enum,
+)
+
+class CHECK(Enum):
+    """An enumeration of the possible results of a check."""
+
+    UNDETERMINABLE = 0
+    TRUE = 1
+    FALSE = 2
